requirements.txt
python-socketio==5.10.0
eventlet==0.33.3
